+++ react/src/base/inputs/Select.stories.js
@@ -4,15 +4,10 @@ import { markdownDocumentationLinkBuilder } from '../../../../storybook-utilitie
 
 export default {
   title: 'Components/Input/Select',
-  decorators: [
-    story => <div className="sprk-o-Box">{story()}</div>
-  ],
+  decorators: [(story) => <div className="sprk-o-Box">{story()}</div>],
   component: SprkSelectionInput,
   parameters: {
-    jest: [
-      'SprkErrorContainer',
-      'SprkInputIconCheck',
-    ],
+    jest: ['SprkErrorContainer', 'SprkInputIconCheck'],
     info: `${markdownDocumentationLinkBuilder('input')}`,
   },
 };
@@ -59,9 +54,7 @@ export const selectBox = () => (
 selectBox.story = {
   name: 'Default',
   parameters: {
-    jest: [
-      'SprkSelectionInput',
-    ]
+    jest: ['SprkSelectionInput'],
   },
 };
 
@@ -109,9 +102,7 @@ export const invalidSelectBox = () => (
 invalidSelectBox.story = {
   name: 'Invalid',
   parameters: {
-    jest: [
-      'SprkSelectionInput',
-    ]
+    jest: ['SprkSelectionInput'],
   },
 };
 
@@ -158,9 +149,150 @@ export const disabledSelectBox = () => (
 disabledSelectBox.story = {
   name: 'Disabled',
   parameters: {
-    jest: [
-      'SprkSelectionInput',
-    ]
+    jest: ['SprkSelectionInput'],
   },
 };
 
+export const hugeSelectBox = () => (
+  <SprkSelectionInput
+    label="Select Box Label"
+    choices={[
+      {
+        label: 'Option 1',
+        value: 'option-1',
+      },
+      {
+        label: 'Option 2',
+        value: 'option-2',
+      },
+      {
+        label: 'Option 3',
+        value: 'option-3',
+      },
+      {
+        label: 'Grouped Options',
+        options: [
+          {
+            label: 'Grouped Option 1',
+            value: 'grouped-option-1',
+          },
+          {
+            label: 'Grouped Option 2',
+            value: 'grouped-option-2',
+          },
+          {
+            label: 'Grouped Option 3',
+            value: 'grouped-option-3',
+          },
+        ],
+      },
+    ]}
+    variant="hugeSelect"
+    name="select"
+    defaultValue=""
+  />
+);
+
+hugeSelectBox.story = {
+  name: 'Huge',
+  parameters: {
+    jest: ['SprkSelectionInput'],
+  },
+};
+
+export const invalidHugeSelectBox = () => (
+  <SprkSelectionInput
+    label="Select Box Label"
+    choices={[
+      {
+        label: 'Option 1',
+        value: 'option-1',
+      },
+      {
+        label: 'Option 2',
+        value: 'option-2',
+      },
+      {
+        label: 'Option 3',
+        value: 'option-3',
+      },
+      {
+        label: 'Grouped Options',
+        options: [
+          {
+            label: 'Grouped Option 1',
+            value: 'grouped-option-1',
+          },
+          {
+            label: 'Grouped Option 2',
+            value: 'grouped-option-2',
+          },
+          {
+            label: 'Grouped Option 3',
+            value: 'grouped-option-3',
+          },
+        ],
+      },
+    ]}
+    variant="hugeSelect"
+    name="select"
+    valid={false}
+    errorMessage="There is an error in the field."
+    defaultValue=""
+  />
+);
+
+invalidHugeSelectBox.story = {
+  name: 'Invalid Huge',
+  parameters: {
+    jest: ['SprkSelectionInput'],
+  },
+};
+
+export const disabledHugeSelectBox = () => (
+  <SprkSelectionInput
+    label="Select Box Label"
+    choices={[
+      {
+        label: 'Option 1',
+        value: 'option-1',
+      },
+      {
+        label: 'Option 2',
+        value: 'option-2',
+      },
+      {
+        label: 'Option 3',
+        value: 'option-3',
+      },
+      {
+        label: 'Grouped Options',
+        options: [
+          {
+            label: 'Grouped Option 1',
+            value: 'grouped-option-1',
+          },
+          {
+            label: 'Grouped Option 2',
+            value: 'grouped-option-2',
+          },
+          {
+            label: 'Grouped Option 3',
+            value: 'grouped-option-3',
+          },
+        ],
+      },
+    ]}
+    variant="hugeSelect"
+    name="select"
+    disabled
+    defaultValue=""
+  />
+);
+
+disabledHugeSelectBox.story = {
+  name: 'Disabled Huge',
+  parameters: {
+    jest: ['SprkSelectionInput'],
+  },
+};
